Cleaned up folder structure and some bugs
The major bugs are all cleared up now so you guys should be able to create your own account and add food items accordingly. I might add a feature to show the user a preview of what the food is before commiting to the food journal but I've spent WAY too much time on this already instead of studying for the tests. Feel free to experiment with the files!